+++ app/Database/Create/SQLCommands.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Solital\Database\Create;
+use Solital\Components\Model\Model;
+use Solital\Database\Create\Create;
+
+class SQLCommands extends Model
+{
+    public function table()
+    {
+        echo "Build your SQL commands in the 'SQLCommands' class\n\n";
+    }
+}